Add jwt protection to /nft/get
