+++ back/package.json
@@ -18,6 +18,7 @@
     },
     "watch": [
       "paths/",
+      "utils/",
       ".env"
     ],
     "env": {
